// src/types/experience.ts
export interface WorkProps {
  id: number;
  company: string;
  affiliation: string;
  period: string;
  position: string[];
  description: string[];
  location: string;
  img?: string;
  imgUrl?: string;
}

export interface FreelancerProps {
  id: number;
  company: string;
  affiliation: string;
  period: string;
  position: string[];
  description: string[];
  img?: string;
  imgUrl?: string;
}